HighScores
Java program that maintains a list of high scores obtained in a game.
The program first asks the user how many scores they want to maintain
and then repeatedly accepts new scores from the user and adds the score
to the list of high scores (in the appropriate position) if it is
higher than any of the existing high scores.